organize update and drawing logic with mode stack and async script
Update and drawing logic no longer happen directly in the App; instead
they're handled by a combination of a mode stack and an async script.

The *mode stack* is a stack of *modes*, each of which represents a unit
of game logic that can update and draw itself, e.g. a dialog, text box
or menu.  They're more commonly known as "game states"; "mode" is just
my own name for them.

Modes cannot push to or pop from the mode stack, so instead they emit
*mode events* that are handled by the *async script* (or just "script"),
which is an async function (i.e. state machine) that defines the
high-level progression logic of the game.  The script spends most of its
time telling the mode stack to update its top-most mode.  If a mode
update emits a mode event, the script can respond by pushing/popping
modes from the mode stack and manipulating shared game data.

The App updates input data and reads the mode stack in order to draw it,
while the script reads the input data and updates the mode stack, but
Rust has no formal way of sharing data between an async function and the
code that polls it.  We work around this by wrapping individual bits of
data with our own home-grown SharedMut smart pointer, which works as
long as App::update and App::draw don't run in parallel.

All shared data that's wrapped up in SharedMuts are shuttled around the
game logic using *contexts*, the most important of which is the
ScriptContext that the script can access directly.  ModeContext and
DrawContext provide borrowed access to the same data to the mode update
and drawing logic respectively, while limiting fields that would
otherwise cause borrow checking problems.