try different numbers for debug battle and debug level up
Trying to find good numbers for the first boss battle.  If the boss has
the same attack and defense as a level 6 player and has a magic
weakness, it should have about 50 times the player's attack worth of HP.
This leads to a boss fight that lasts a little over 20 turns, uses
healing and attkck magic constantly, and requires the use of a salve and
a tonic to win.

Player max MP growth dropped from 2 to 1; max MP and magic costs dictate
the length of boss battles, and having too much would make those battles
drag on too long.

Attack and defense bumped up to simulate the effect of weapon and armor
upgrades; real numbers for weapon attack and armor defense still need to
be decided.