fill in next level experience table
The gist of the values is for each one to represent "battles times
average experience gained per battle versus an encounter group".  The
game's random encounters are divided into six encounter groups whose
experience values are averaged.  The number of desired battles for each
new level roughly follows the formula: `3 * (ln(level / 10) + 3)`.

Levels 5, 9, 13, 17 and 20 use the average experience per battle of the
previous and current encounter group to smooth transitions between
levels a bit.

Levels beyond 24 are just +20% compounded, leading to values that are
large but reachable with patience.