refactor data sharing between game loop and async script function
The main `App` struct no longer holds a `ScriptContext` clone; this
means that the shared `input`, `modes`, `level`, `actors` and `fade`
fields are held directly by the `App` struct.

The `rng`, `progress` and `steps` fields of `ScriptContext` never needed
to be wrapped with `SharedMut`, so now they're not.

The `input`, `level` and `fade` fields of `ModeContext` didn't need to
be mutable, so now they're not.

The `ScriptContext` struct no longer needs a `clone` method, so it's
been removed; this means that the `Resources` struct no longer needs to
implement the `Clone` trait, which in turn means that its `levels`,
`textures_by_path` and `sprite_sheets_by_path` fields no longer need to
be wrapped with `Rc`.