add ability to save and load progress
The player can now choose to save the game after using the bed.  If
saved progress is detected on the title screen, a choice to "Continue"
will appear that loads this progress when selected.

Using quad-storage, saved data will be written to a "local.data" file in
native builds and to LocalStorage when running in WebAssembly.