add chance for enemy turn when choosing to run from battle
When the player chooses to run from battle, the enemy may get a turn
before the player gets away.  The chance for the player to avoid this is
based on the total damage taken by the enemy plus whatever the player
would have inflicted had they chosen to fight instead; higher total
damage reduces the chance of an enemy turn, and if the player would have
defeated the enemy it'll never get that turn.