change keys: space to confirm, left ctrl to cancel/menu, '\' to debug
The 'z' and 'x' keys may not be next to each other on non-US keyboards.
Also, the debug menu key was too close to them, so it's been moved to
the backslash key instead.

I would have used Esc for cancel/menu, but it might already be used by
the browser in the WebAssembly build, so I avoided it.