play music throughout the game
This adds background music that will play throughout most of the game,
except for the title screen.

While walking around, background music that has been set for the current
level will be played.

If a battle occurs, either normal battle music or a boss battle theme
will play for the duration of the battle.

The introduction and ending sequences both have unique music tracks that
will play while they run.

The WebAssembly build has to pause for a tiny moment for music to load
before being played, which isn't mentioned in the docs for quad-snd.