add wasm-site, wasm-serve and wasm-zip tasks
`wasm-site` builds the WebAssembly build of the game in release mode and
copies the `.wasm` file to the `$WASM_DIR` directory.

`wasm-serve` runs a test HTTP server in the `$WASM_DIR` directory so the
website with the WebAssembly build can be tested in a browser.

`wasm-zip` ensures the `$WASM_DIR` directory has the most recent
WebAssembly build and adds it and all the website files to a zip file as
top-level file entries.

The `WASM_DIR` environment variable must be set to the website directory
(checking for the existence of `gl.js` there) for all of these tasks.